hack: Support version in build-release.sh
This change introduces a version file and uses it in the
build-release.sh script. The version compiled into the kf binary is
changed to "semver(-nightly) git-sha", where nightly is appended if
$NIGHTLY is set.

Version is also replaced into the K8s manifests used by Ko.
Specifically, the version is set on the label of the kf namespace.